Update config.yaml to correct Huanyang spindle config
Corrected tx pin to be 4, rx pin to be 25, (they were opposite and incorrect). Also corrected overall config adding uart1 as stand alone along with Huanyang and uart_num: 1 which was missing. Please note, the data between the comments labeled start and end is what you need for this to run on XPro V5.